+++ src/services/axios/post/postProposal/mockProposalBackend.tsx
@@ -22,6 +22,7 @@ export const MockProposalBackend: ProposalBackend = {
     observation_sets: [],
     data_product_sdps: [],
     data_product_src_nets: [],
-    result_details: []
+    result_details: [],
+    calibration_strategy: []
   }
 };