+++ CHANGELOG.md
@@ -11,6 +11,10 @@ UNRELEASED
 * [BREAKING]: Changed: PDM changes implemented
 * [BREAKING]: Added calibration to proposal types and mappings
 * Changed: oso-services v9.0.0
+* Changed: Implemented the isSV function so that specific SV flow thru the application can be realized.
+* Changed: oda v12.1.1
+* Added: polarisations to mapping
+* Fixed: use of REACT_APP_USE_LOCAL_DATA for local development
 * Changed: to show notification when there is an senscalc error after mapping target-observation
 
 8.0.0